Refactor import paths in App and ColorInputGroup components for better organization. Move getContrastColor utility function to a separate file. Update ColorMixer component to improve code readability with block scoping in the weighted mixing case. Enhance MixHistoryItem component by importing React explicitly.
