+++ src/App.tsx
@@ -3,7 +3,7 @@ import './App.css'
 import Header from './components/Header'
 import ColorMixer from './components/ColorMixer/index'
 import ColorPalette from './components/ColorPalette'
-import MixHistory from './components/MixHistory'
+import MixHistory from './components/MixHistory/index'
 import { MixRecord } from './components/MixHistory/types'
 import { MixingAlgorithm } from './components/ColorMixer/ControlPanel'
 import { v4 as uuidv4 } from 'uuid'
