+++ src/components/ColorMixer/ColorInputGroup.tsx
@@ -1,5 +1,6 @@
 import { FC } from 'react';
 import { RandomColorIcon } from '../icons';
+import { getContrastColor } from './utils';
 
 export interface ColorInputGroupProps {
   color: string;
@@ -10,18 +11,6 @@ export interface ColorInputGroupProps {
   emptyColor: string;
   readOnly?: boolean;
 }
-
-// Utility function to determine text color based on background
-export const getContrastColor = (hexColor: string) => {
-  const r = parseInt(hexColor.substr(1, 2), 16);
-  const g = parseInt(hexColor.substr(3, 2), 16);
-  const b = parseInt(hexColor.substr(5, 2), 16);
-
-  // Calculate brightness (weighted RGB)
-  const brightness = (r * 299 + g * 587 + b * 114) / 1000;
-
-  return brightness > 128 ? '#000000' : '#FFFFFF';
-};
 
 export const ColorInputGroup: FC<ColorInputGroupProps> = ({
   color,
